Adding Date Filter to the logs GET call
If the FromDate parameter is added, it will now be used to filter the logs
that are sent back. By default this is a date that is far back (1-1-0001)
so if not set, it will effectively return without filtering on date.

In both cases though, note that only the latest record for a particular
board and user combination is sent back.